Implement the ability to specify the position and size of textures (in pixels) The 'vertices' array, which is defined in the 'Core::start_main_loop' function, stores the local coordinates of the vertices of two triangles (in the range [0..1]) and the corresponding texture coordinates. The coordinates of each vertex are passed to the 'in_pos' vector of the vertex shader (the order is determined by the contents of GL_ELEMENT_ARRAY_BUFFER, i.e. the 'indices' array). Also, uniforms vec2 'uf_model_pos' and vec2 'uf_model_size' are passed to the vertex shader. These variables set the position of the vertices (in pixels) and the size of the primitives formed from these vertices (in pixels), respectively. Position and size are set as follows: A 4x4 'model' matrix is created in the vertex shader. Scaling along the x and y axes is implemented by changing the first two elements of the main diagonal of this matrix (ie 'model[0][0]', 'model[1][1]'). Shifting along the x and y axes is implemented by changing the elements with indices [3][0], [3][1], respectively. Thus, scaling along the x and y axes in the vertex shader is implemented in the lines: model[0][0] = uf_model_size.x; model[1][1] = uf_model_size.y; , and the position change is implemented in the lines: model[3][0] = uf_model_pos.x; model[3][1] = uf_model_pos.y; To avoid distortions in further transformations, the remaining values of the main diagonal of the 'model' matrix (i.e., 'model[2][2]', 'model[3][3]') should be filled with ones, and everything else (except elements used for scaling and shifting) must be filled with zeros. The result of multiplying this matrix by the vector 'in_pos' is a coordinate vector in the world space coordinate system.
In addition, the projection matrix (uniform variable 'uf_projection') is passed to the vertex shader. The projection matrix is created based on the window size by calling the 'glm::ortho' function and passed to the shader before the start of the render loop. Multiplication involving the projection matrix converts the coordinates given in pixels to normalized ones, that is, it "projects" the scene onto the working area of the window (i.e. the area specified by the 'glViewport' function).

To demonstrate the implemented capabilities, 2 textures were created (in fact, these are 2 layers of an array of 2d textures). Inside the render loop, before each layer is drawn, the 'uf_model_pos' and 'uf_model_size' values are sent to the shader (size 512x512 pixels, position (0,0) for the first layer, size 256x128 pixels, position (512,0) for the second layer). As a result, two textures are drawn side by side.
--------
The 'Core::window_size_' field is initialized in the 'Core::init window' function.
--------
Changed the version of the vertex and fragment shaders to 4.3.