Create 'VertexArray' class This class implements logic for creating, filling with data, building and freeing vertex arrays.
Creation of a vertex array is implemented in the constructor of the class and consists in generating a vertex array object and related buffer objects needed to store renderable data (vertex buffer, texture vertices buffer, indices buffer).

Filling a vertex array with data is implemented in the 'add_textured_rect' method and consists in adding vertices to the data required to build the vertex array and generating an indices array corresponding to this vertices.

Building of a vertex array is:
- filling in OpenGL buffer objects (vertices, texture vertices and indices) with data (i.e., sending this data to the GPU);
- binding these buffer objects to the vertex array.

Freeing a vertex array is implemented in the destructor of the class and consists in unbinding and deleting the vertex array object and related buffers. According to the OpenGL documentation, deleting a currently bound buffer unbinds it implicitly.