Draw multiple textured rectangles using one 'glDrawElements' call
The functionality that creates and initializes a vertex array has been removed from the 'start_main_loop' function. Now an object of the 'VertexArray' class is used to create and fill a vertex array.

A vertices of 1 more textured rectangle has been added to a vertex array (the second rectangle is located on the first rectangle and is 2 times smaller). To draw the second rectangle, inside the render loop, in one of the calls to the 'glDrawElements' function, the second argument (number of elements to be rendered) has been changed from 6 to 12.

The logic for deleting a vertex array object and associated vertex buffers has been removed from the 'start_main_loop' function, because this is now implemented in the destructor of the 'VertexArray' class.

Thus, one 'glDrawElements' call renders two textures. In the future, this feature will be used to render static scenes with a large number of different textures.