Use the 'IndicesData' class to store textured rectangles indices data
Before starting the render loop, two textured compositions are initialized:
 - The first composition is described by arrays 'vertices_1' (vertices) and 'txd_vertices_1' (textured vertices and is a textured rectangle.
 - The second composition is described by arrays 'vertices_2' and 'txd_vertices_2' and consists of two textured rectangles partially superimposed on each other.

Vertices and texture vertices of each composition are passed to a vertex array (by calling the 'VertexArray::add_textured_rects' function).
The function places these for further sending to the GPU and returns an object of the 'IndicesData' class for each composition ('indices_data_1', 'indices_data_2').

Inside the render loop, compositions are rendered by calling the 'glDrawElements' function. The arguments for this function are taken from 'indices_data_1' and 'indices_data_2' respectively.